backend/requirements.txt
# Core API
fastapi>=0.115,<1.0
uvicorn[standard]>=0.28,<1.0

# Uploads
python-multipart>=0.0.9

# Database (PostgreSQL)
sqlalchemy>=2.0
psycopg2-binary>=2.9

# (Firebase removed)

# Storage / external services
boto3>=1.34
requests>=2.32

# Images & EXIF
pillow>=10.0
exifread>=3.0

# Background jobs
redis>=5.0
rq>=1.15

# Config
python-dotenv>=1.0
